Add NeedsExternalResource JUnit Category * Such tests are being excluded on Travis. * Converted TestClientCatalogBasic to ClientCatalogBasicSpec, because Spock has better parameterization. We needed to apply the category to only some of the parameters.
