+++ cdm/build.gradle
@@ -99,6 +99,7 @@ task testJar(type: Jar, dependsOn: testClasses,
         // JUnit categories
         include 'ucar/unidata/test/util/NeedsCdmUnitTest.class'
         include 'ucar/unidata/test/util/NeedsContentRoot.class'
+        include 'ucar/unidata/test/util/NeedsExternalResource.class'
         include 'ucar/unidata/test/util/NotJenkins.class'
         include 'ucar/unidata/test/util/NotTravis.class'
     }
