+++ cdm/src/test/java/ucar/unidata/test/util/NeedsExternalResource.java
@@ -1,0 +1,11 @@
+package ucar.unidata.test.util;
+
+/**
+ * A marker to be used with JUnit categories to indicate that a test method or test class requires access to an
+ * external resource in order to complete successfully. Most commonly, these are remote THREDDS servers.
+ *
+ * @author cwardgar
+ * @since 2015-10-12
+ */
+public interface NeedsExternalResource {
+}
